Add path truncation, git SHA to prompt
Pulled in truncation code from previous zsh prompt to keep the path to a
single line on screen. In doing so, I had to add a pile of bare_
functions to produce an uncolored version of the non-path portions of
the prompt (user@host, git info) for purposes of determining the length
available for the path.

Heavily divergent from blinks at this point, enough that there's hardly
even a whiff of its code left. I've swapped out all the nonfunctioning
color codes (%B%F nonsense) for some that actually work ($fg[color]).