Add custom directory to git
It may make sense to have custom in .gitignore if your goal is to keep
the oh-my-zsh repo clean, but it's utterly useless if you're working
with your own fork and you want to share customizations between multiple
machines. Took me a few days to figure out why my key bindings worked on
one machine but not others.